Introduced new "SkyfishHttpHelper" class
Normally I leave it up to the packages or projects using the integration package to "interpret" the underlying API, but for the sake of this example, this could also partially be handled by a new helper class.

For one, the helper class contains the "GetVideoByMediaId" and "GetVideoByUniqueMediaId" methods. These don't really exist in the underlying API, so they are really just calling the "Search" method with the right parameters, delegating the work to the underlying parts of the package.